Adding the 'shared' keyword.
This keyword behaves similarly to the 'case' keyword, in that it creates a
package nested under the current namespace, but it does not create a
subclass of Test::Mini::TestCase.  Instead, it creates a package which will
export all of its test methods to the calling package.  This also bundles
the setup and teardown advice declared in the shared example.

The biggest way in which this differs from the behavior of the 'case'
keyword is in the fact that shared examples may not be nested.